requirements.txt
tensorflow==1.15.2
nltk==3.5
Keras==2.2.4
opencv_python==3.4.3.18
requests==2.24.0
scipy==1.1.0
matplotlib==3.0.0
h5py==2.8.0
numpy==1.19.1
pandas==1.1.0
imutils==0.5.1
protobuf==3.13.0
scikit_learn==0.23.2
google-cloud-storage